PeppulelluBot_Workspace_AUTO_STREAM/requirements.txt
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.61.0
python-dotenv==0.20.0